Agrego integración de Mercado Pago. Se implementan variables de entorno que no se suben al repositorio, obviamente la versión de desarrollo. Se relanza el conteiner del docker para permitir transacciones.
Agrego código para SUPERADMIN.
Permito que el admin vea las orders pagas (statusHistory = "PAID").

+++ Backend/API Express/src/LineOrder/lineOrder.controller.ts
@@ -82,7 +82,11 @@ async function remove(req: Request, res: Response) {
   try {
     const id = req.params.id
     const lineOrder = em.getReference(LineOrder, id)
+    if (!lineOrder) {
+      return res.status(404).json({ message: "lineOrder not found" });
+    }
     await em.removeAndFlush(lineOrder)
+    return res.status(200).json({ message: 'lineOrder removed' })
   } catch (error: any) {
     res.status(500).json({ message: error.message })
   }
